Extract comments
