+++ front/app/containers/Admin/inspirationHub/ProjectDrawer/index.tsx
@@ -1,16 +1,7 @@
 import React from 'react';
 
-import {
-  Box,
-  Text,
-  Title,
-  Button,
-  colors,
-  fontSizes,
-  stylingConsts,
-} from '@citizenlab/cl2-component-library';
+import { Box, Text, colors } from '@citizenlab/cl2-component-library';
 import { useSearchParams } from 'react-router-dom';
-import styled from 'styled-components';
 
 import useAppConfiguration from 'api/app_configuration/useAppConfiguration';
 import useProjectLibraryProject from 'api/project_library_projects/useProjectLibraryProject';
@@ -24,39 +15,10 @@ import { removeSearchParams } from 'utils/cl-router/removeSearchParams';
 
 import { useLocalizeProjectLibrary } from '../utils';
 
-import Comment from './Comment';
+import Comments from './Comments';
 import Header from './Header';
 import messages from './messages';
 import Phase from './Phase';
-
-const StyledBox = styled(Box)`
-  font-size: ${fontSizes.base}px;
-  height: 40px;
-
-  &:focus {
-    outline: 1px solid ${colors.primary};
-    height: auto;
-  }
-`;
-
-const DUMMY_DATA = [
-  {
-    name: 'Michael Bluth',
-    createdAt: '2021-08-10T12:00:00Z',
-    badgeText: 'Go Vocal',
-    badgeType: 'go-vocal',
-  },
-  {
-    name: 'Tobias Funke',
-    createdAt: '2021-08-10T12:00:00Z',
-    badgeText: 'Vienna',
-    badgeType: 'platform-moderator',
-  },
-  {
-    name: 'George Michael Bluth',
-    createdAt: '2021-08-10T12:00:00Z',
-  },
-] as const;
 
 const ProjectDrawer = () => {
   const [searchParams] = useSearchParams();
@@ -118,27 +80,7 @@ const ProjectDrawer = () => {
             ))}
           </Box>
           <Box mt="32px">
-            <Title variant="h3">Comments</Title>
-            <StyledBox
-              as="textarea"
-              rows={5}
-              width="100%"
-              maxWidth="500px"
-              border={`1px solid ${colors.borderDark}`}
-              borderRadius={stylingConsts.borderRadius}
-              p="10px"
-              placeholder="Write your comment here"
-            />
-            <Box w="100%" mt="8px" display="flex">
-              <Button w="auto" bgColor={colors.primary}>
-                Post your comment
-              </Button>
-            </Box>
-            <Box mt="20px">
-              {DUMMY_DATA.map((comment, i) => (
-                <Comment {...comment} key={i} />
-              ))}
-            </Box>
+            <Comments />
           </Box>
         </Box>
       )}
